build(bench): add "consume" and "echo" benchmarks
This fleshes out benchmarks with speed tests of sending bytes into
Wasm linear memory from an input and copying those bytes back out.
This way we can get a sense of the performance of those operations
and how they relate to the full roundtrip in "reflect".

The source of these plugins is available in `extism/plugins` as
`echo.wat` and `consume.wat` [1].

[1]: https://github.com/extism/plugins/pull/4/files