Update wasmtime version with exceptions support
The wasmtime v37 release[1] "now fully implements the WebAssembly
exception-handling proposal." Support for exception-handling is disabled
by default, but will be made default in the future (according to the
release notes).

To use this new wasmtime feature, we do two things at a high level:

	1. Update wasmtime dependency to version 37
	2. Add a "wasmtime-exceptions" Extism feature, disabled by
	   default.

Updating to the new wasmtime dependency required some semantical changes
to how the cache is configured.

There is also a new `ExternType` to match on when handling invalid
imports.

[1] https://github.com/bytecodealliance/wasmtime/releases/tag/v37.0.0